Fix: native events for component parts with the same name were not properly canceled
Previously, due to unthoughtful implementation, native events for component parts
that had the same name were not all canceled: if you had three html elements with
`data-component-part="part3"` attribute and native events defined for "self.part3"
in the component, then after calling `Component._cancelNativeEvents()`
only the first html element's events would be be canceled.

This also fixes the bug with SelectComponent empty options list, which was the direct
result of this issue.