DisplayStateManager applies latest state only, ignoring all the previous ones in the queue
The purpose of this method is to always attempt to apply the latest changes, ignoring
all the previous changes whose line in the queue may not have yet come. In fact, there's no queue.

Say there are 3 state changes: (a) is already being animated, (b) is sitting waiting its turn in
`this.last_state_change` property and (c) was just initiated and applyChanges() just called for it.
In this case (b) will be comletely ingored, as applyChanges() will rewrite `this.last_state_change`
and put (c) into it. Therefore, after the animation for (a) is completed, DisplayStateManager will get
straight to executing (c) animations. This is because there's no point in animating (b) because the state
had already changed since then.